Restructure for Apollo links
